+++ Assets/Scripts/UI/Leaderboard.cs
@@ -18,14 +18,14 @@ namespace EcoBuilder.UI
             {
                 lockShade.SetActive(false);
             }
-            int playerScore = GameManager.Instance.GetHighScoreLocal(levelPrefab.Details.Idx);
+            long playerScore = GameManager.Instance.GetHighScoreLocal(levelPrefab.Details.Idx);
             nameText.text = "loading high scores...\n\n\nyour score";
             scoreText.text = "\n\n\n" + playerScore;
             return Instantiate(levelPrefab, levelParent);
         }
         public void SetFromGameManagerCache()
         {
-            int median = GameManager.Instance.GetLeaderboardMedian(LevelToPlay.Details.Idx);
+            long median = GameManager.Instance.GetLeaderboardMedian(LevelToPlay.Details.Idx);
             var scores = GameManager.Instance.GetLeaderboardScores(LevelToPlay.Details.Idx);
             if (median == -1) {
                 return;
@@ -38,7 +38,7 @@ namespace EcoBuilder.UI
                 scoreText.text += (scores.Count>i? scores[i].Item2.ToString() : "") + "\n";
             }
             nameText.text += "you\nworld average";
-            int playerScore = GameManager.Instance.GetHighScoreLocal(LevelToPlay.Details.Idx);
+            long playerScore = GameManager.Instance.GetHighScoreLocal(LevelToPlay.Details.Idx);
             scoreText.text += playerScore + "\n" + median;
         }
     }
